requirements.txt
django
pillow
django-admin-interface
django-crispy-forms
crispy-bootstrap5
mysqlclient
pymysql
djangorestframework